Implement wavelet spline baseline removal
