+++ point_spectra_gui/core/baselineRemovalMethods/__init__.py
@@ -1,6 +1,6 @@
 from point_spectra_gui.core.baselineRemovalMethods import ALS
 from point_spectra_gui.core.baselineRemovalMethods import AirPLS
-from point_spectra_gui.core.baselineRemovalMethods import SWT
+from point_spectra_gui.core.baselineRemovalMethods import WaveletSpline
 from point_spectra_gui.core.baselineRemovalMethods import Dietrich
 from point_spectra_gui.core.baselineRemovalMethods import FABC
 from point_spectra_gui.core.baselineRemovalMethods import KK
@@ -9,3 +9,5 @@ from point_spectra_gui.core.baselineRemovalMethods import Mario
 from point_spectra_gui.core.baselineRemovalMethods import Median
 from point_spectra_gui.core.baselineRemovalMethods import Polyfit
 from point_spectra_gui.core.baselineRemovalMethods import Rubberband
+
+
